Finished folder refactoring...
