+++ fits_reduce/tests/test_config.py
