+++ fits_reduce/util/slices.py
